src/builder/requirements.txt
grpcio-tools ~= 1.30
google-cloud-build